Conflicts:
	src/cf-serverd.c

+++ src/conversion.c
@@ -22,13 +22,14 @@
   included file COSL.txt.
 */
 
-#include "cf3.defs.h"
+#include "conversion.h"
 
 #include "promises.h"
 #include "files_names.h"
 #include "dbm_api.h"
 #include "mod_access.h"
 #include "item_lib.h"
+#include "reporting.h"
 
 static int IsSpace(char *remainder);
 
@@ -1322,7 +1323,7 @@ static int IsSpace(char *remainder)
 
     for (sp = remainder; *sp != '\0'; sp++)
     {
-        if (!isspace(*sp))
+        if (!isspace((int)*sp))
         {
             return false;
         }
